requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
SQLAlchemy==2.0.23
nltk==3.8.1
sendgrid==6.10.0
python-dotenv==0.21.1
PyYAML==6.0.1
